# [IMP] stock: improve stock_inventory action_start perfs