Backport branch odoo/odoo#79587 to v13.

In `_get_inventory_lines_values`, remove quants recordset
as it has been unused since commit cc9324f74084.

Browse product.product in `_get_inventory_lines_values`
outside of for loop to speedup getting the uom_id.


#### speedup

Time reported = calling `inventory.action_start` on an inventory with 10.000 products.

| Before PR | Backport only | After PR |
|:----------:|:--------------:|:---------:|
|  2min | 1min24s | 9.30s |


_NB : This branch should be ignored by fw-bot._

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr