@Yenthe666 2021-11-19
You sir, deserve a 🥇! Thank you from the Odoo users for backporting this :)