@Aurelienvd 2021-12-03
@odoo/rd-logistics Hello, could this PR be r+ since it has already been positively reviewed ?